requirements.txt
python-telegram-bot==20.7
openai>=1.40.0
gspread==6.1.2
google-auth==2.31.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
